Fixes Model utils
